created Strain app and made new migrations
